Get available indices

This functions returns a dataframe with the available indices for calcualtion. Use the values found in the column name to specify an individual index

check_indices()

Value

A dataframe

Author

Darius Görgen (MapTailor Geospatial Consulting GbR) info@maptailor.net
Maintainer: MAPME-Initiative contact@mapme-initiative.org
Contact Person: Dr. Johannes Schielein
Copyright: MAPME-Initiative
License: GPL-3

Examples

library(mapme.vegetation)
check_indices()[1:10,]
#>                                        longname      name
#> 1            Aerosol free vegetation index 1600  AFRI1600
#> 2            Aerosol free vegetation index 2100  AFRI2100
#> 3                 Anthocyanin reflectance index       ARI
#> 4                      Ashburn Vegetation Index       AVI
#> 5  Atmospherically Resistant Vegetation Index 2     ARVI2
#> 6      Blue-wide dynamic range vegetation index    BWDRVI
#> 7                    Browning Reflectance Index       BRI
#> 8                       Chlorophyll Index Green   CIgreen
#> 9                      Chlorophyll IndexRedEdge CIrededge
#> 10                             Coloration Index        CI
#>                                                  link
#> 1  http://www.indexdatabase.de/db/i-single.php?id=393
#> 2  http://www.indexdatabase.de/db/i-single.php?id=395
#> 3  http://www.indexdatabase.de/db/i-single.php?id=214
#> 4  http://www.indexdatabase.de/db/i-single.php?id=574
#> 5  http://www.indexdatabase.de/db/i-single.php?id=396
#> 6  http://www.indexdatabase.de/db/i-single.php?id=136
#> 7  http://www.indexdatabase.de/db/i-single.php?id=480
#> 8  http://www.indexdatabase.de/db/i-single.php?id=128
#> 9  http://www.indexdatabase.de/db/i-single.php?id=131
#> 10  http://www.indexdatabase.de/db/i-single.php?id=11
#>                                       s2_formula
#> 1  (band_8-0.66*(band_11)/(band_8+0.66*band_11))
#> 2   (band_8-0.5*(band_12)/(band_8+0.56*band_12))
#> 3                      (1)/(band_3)-(1)/(band_5)
#> 4                              2.0*band_9-band_4
#> 5   -0.18+1.17*((band_8-band_4)/(band_8+band_4))
#> 6        (0.1*band_8-band_2)/(0.1*band_8+band_2)
#> 7           ((1)/(band_3)-(1)/(band_5))/(band_8)
#> 8                            (band_8)/(band_3)-1
#> 9                            (band_8)/(band_5)-1
#> 10                      (band_4-band_2)/(band_4)