Calculate a spectral index with gdalcubes

This function is used internally to calculate a spectral index using the gdalcubes package. It is called by calcIndex() on a given set of SAFE directories. A user defined cube view defined by the x and y resolution, the time resolution, a CRS, as well as aggregation methods is applied. Additionally, the time dimension can be aggregated which is useful for seasonal calculations of a given index. The output of the function is a GTiff file for each time step labeled as <label>_<index>_<timestep>.tif. Zonal statistics are written to a Geopackage for each timestep labeled as <label>_<index>_<timestep>.gpkg to outdir.

cube_indexcalc(
  safes,
  aoi,
  epsg = NULL,
  dx = NULL,
  dy = NULL,
  dt = NULL,
  aggregation = "mean",
  resampling = "bilinear",
  tmpagg = FALSE,
  index = NULL,
  label = "",
  outdir = ".",
  threads
)

Arguments

safes
A character vector with the .jp2 and .tif files of the selected SAFE directories. Obtained with list.files(<safedir>, "jp2$|.tif$", recursive =T, full.names = T)

aoi
A sf-object with POLYGONS or MULTIPOLYGONS. For objects with ZM dimensions these dimensions are dropped before calculation.

epsg
A character specifying the epsg code for the the cube view specified as "EPSG:<code>". The resulting GTiff files and the GPKG with the zonal statistics will be projected into this CRS. dx and dy are understood in the units of the CRS specified here. If left empty, the CRS of the aoi variable will be used.

dx
A numeric specifying the resolution of the output raster in x dimension expressed in the map unit specified in epsg. If epsg is left empty it defaults to the map unit of aoi.

dy
A numeric specifying the resolution of the output raster in y dimension expressed in the map unit specified in epsg. If epsg is left empty it defaults to the map unit of aoi.

dt
Resolution of the time dimension expressed as ISO8601 string (e.g. "P16D" for 16 days).

aggregation
A character vector specifying the aggregation method for images of the same time stamp. Must be one of "min", "max", "mean", "median", or "first". Defaults to "mean".

resampling
A character specifying the resampling method used to transform images from their native CRS to the output CRS. Can be any of the methods supported by gdalwarp (see www.gdal.org/programs/gdalwarp.html). Defaults to "bilinear".

tmpagg
A logical indicating if temporal aggregation of all timestemps should be performed calling reduce_time(). Defaults to FALSE.

index
A character specifying the short name of the index to be calculated. Available indices can be checked with check_indices().

label
A character label which is appended to the output files in the form of <label>_<index>_<timestep>. Files with the same name in outdir will be overwritten without warning.

outdir
A character specifying an existing directory where output files will be written to. Defaults to ".".

threads
Number of cores used for parallel processing

zonalstat
A character specifying the aggregation statistic to be used by zonal_statistics(). Multiple selections are supported. Must be included in "min", "max", "mean", "median", "count", "sum", "prod", "var", and "sd". Defaults to "mean".

Value

Nothing. However a GTiff and a GPKG with the zonal statistics are written to outdir for every resulting timestep.

Note

See cube_view() and zonal_statistics() for more information.

Author

Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) info@maptailor.net
Maintainer: MAPME-Initiative contact@mapme-initiative.org
Contact Person: Dr. Johannes Schielein
Copyright: MAPME-Initiative
License: GPL-3