Download items from a STAC catalog

This function is used internally to download a given list of items matching a spatiotemporal query.

.download_assets(
  items = NULL,
  assets = NULL,
  outdir = ".",
  use_aria = FALSE,
  aria_bin = NULL,
  skip_existing = FALSE,
  max_concurrent = 4,
  max_connections = 4,
  retry = 2,
  verbose = TRUE
)

Arguments

items
A list of items returned by `rstac::stac_search()`

assets
Default is NULL. A possible charachter vector to specify which assets to download. If assets do not match the function returns an error and lists the available assets.

outdir
A character vector of an existing directory. Defaults to the current working directory.

use_aria
A logical if the aria2 downloader shall be used. Defaultes to FALSE which will download the items sequentially via GET requests. If TRUE it is expected that the argument `aria_bin` points towards a valid aria binary installation.

aria_bin
A charachter vector pointing to a valid aria binary.

skip_existing
A logical indicating if existing files should be removed from the download list prior to downloading with aria. Might speed up the process when a prior download was interrupted, however, when files were not downloaded completley their download is not ressumed (check for .aria files in outdir). Defaults to FALSE which means that existing files are not skipped.

max_concurrent
Maximum number of connections to the server to load a single file. Defaults to 4.

max_connections
Maximum number of parallel connections to the server to load files simoltaniously. Defaults to 4.

retry
Integer. Maximum integer of retries for erroneous downloads.

verbose
A logical indicating the verbosity.

Value

Nothing. Downloads specified assets into `outdir`.

Author

Darius Görgen (MapTailor Geospatial Consulting GbR) info@maptailor.net
Maintainer: MAPME-Initiative contact@mapme-initiative.org
Contact Person: Dr. Johannes Schielein
Copyright: MAPME-Initiative
License: GPL-3